requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pdfminer.six==20231228